requirements.txt
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap-modal-forms==1.5.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-tables2==2.3.1
feedparser==5.2.1
gunicorn==20.0.4
idna==2.9
mysqlclient==1.4.6
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
requests-file==1.5.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tldextract==2.2.2
urllib3==1.25.9
whitenoise==5.2.0
